Renames button component
